+++ test/exis/xrsfl2flsum_simple/test_xrsfl2flsum.py
@@ -4,6 +4,7 @@ from ncagg.config import Config
 from ncagg.aggregator import generate_aggregation_list, evaluate_aggregation_list
 import glob
 import os
+import netCDF4 as nc
 
 class TestAggregate(unittest.TestCase):
     def setUp(self):
@@ -18,17 +19,18 @@ class TestAggregate(unittest.TestCase):
 
     def test_main(self):
         """
-        Just make sure this completes without error.
-
         Nothing too fancy here, but just making sure that aggregating
         a variable of strings works properly.
 
-        Previously, we've had trouble with vlen datatypes.
+        Previous to version 0.8.5 we had trouble with vlen datatypes.
         """
         agg_list = generate_aggregation_list(self.config, self.files)
         evaluate_aggregation_list(self.config, agg_list, self.file)
-
-        # TODO: look at the output _FIllValue for the vlen datatype
-        # make sure that's correct.
+        
+        with nc.Dataset(self.file) as nc_in:
+            status = nc_in.variables["status"]
+            # there should be no fill values...
+            # before ncagg v0.8.5 vlen types like string incorrectly aggregated to all fill values.
+            self.assertFalse(any(status[:] == status._FillValue))